Fix int64 style, ts_func.cpp overflow, and test conventions
- Replace long long with OpenCV-conventional int64 type
- Fix same overflow bug in modules/ts/src/ts_func.cpp duplicate
- Align test code style with OpenCV conventions

+++ modules/core/src/copy.cpp
@@ -938,11 +938,8 @@ int cv::borderInterpolate( int p, int len, int borderType )
     {
         CV_Assert(len > 0);
         if( p < 0 )
-            // Use long long arithmetic to avoid signed integer overflow when p is
-            // close to INT_MIN. The sub-expression (p - len + 1) can underflow if
-            // performed in 32-bit, which is undefined behavior.  The final result
-            // is always in [0, len) and fits safely back in int.
-            p = (int)((long long)p - (((long long)p - len + 1) / len) * (long long)len);
+            // Use int64 to avoid signed integer overflow when p is close to INT_MIN
+            p = (int)((int64)p - (((int64)p - len + 1) / len) * (int64)len);
         if( p >= len )
             p %= len;
     }
